backend/requirements.txt
openai>=1.0.0,<2.0.0
fastapi>=0.110.0,<1.0.0
uvicorn>=0.29.0,<1.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0,<2.0.0
pydub>=0.25.1,<1.0.0
PyJWT>=2.8.0,<3.0.0
bcrypt>=4.1.0,<5.0.0
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<1.0.0
